+++ frontend/src/components/character/CharacterList.vue
