+++ frontend/src/components/character/CharacterTabs/SkinsTab.vue
