+++ frontend/src/components/character/CharacterTabs/UpgradeTab.vue
