+++ frontend/src/components/quests/QuestCharacter.vue
