+++ frontend/src/components/quests/QuestComponentIcon.vue
