+++ frontend/src/components/quests/QuestRequirements.vue
@@ -32,7 +32,7 @@
 import Vue from 'vue';
 import {PropType} from 'vue/types/options';
 import {Quest, QuestItemsInfo, Rarity, RequirementType, RewardType, DustRarity} from '@/views/Quests.vue';
-import QuestComponentIcon from './QuestComponentIcon.vue';
+import QuestComponentIcon from '../quests/QuestComponentIcon.vue';
 import {questItemTypeSupportsStars, questItemTypeSupportsTimesValue} from '@/utils/common';
 import {mapActions} from 'vuex';
 import questItemsInfo from '@/data/questItems.json';
