+++ frontend/src/components/quests/QuestSubmissionModal.vue
