+++ frontend/src/views/Character.vue
@@ -157,9 +157,9 @@ import i18n from '@/i18n';
 import { Nft } from '@/interfaces/Nft';
 import { CharacterPower } from '@/interfaces';
 import BigButton from '@/components/BigButton.vue';
-import CharacterList from '@/components/smart/CharacterList.vue';
+import CharacterList from '@/components/character/CharacterList.vue';
 import CharacterNav from '@/components/layout/navigation/CharacterNav.vue';
-import Character from '@/components/smart/Character.vue';
+import Character from '@/components/character/Character.vue';
 import Events from '@/events';
 import { getConfigValue } from '@/contracts';
