+++ frontend/src/views/Quests.vue
@@ -124,10 +124,10 @@ import Vue from 'vue';
 import {mapActions, mapGetters, mapState} from 'vuex';
 import {Nft} from '@/interfaces/Nft';
 import {Accessors} from 'vue/types/options';
-import QuestRow from '@/components/smart/QuestRow.vue';
-import QuestComponentIcon from '@/components/smart/QuestComponentIcon.vue';
+import QuestRow from '@/components/quests/QuestRow.vue';
+import QuestComponentIcon from '@/components/quests/QuestComponentIcon.vue';
 import QuestReward from '@/components/smart/QuestReward.vue';
-import QuestsList from '@/components/smart/QuestsList.vue';
+import QuestsList from '@/components/quests/QuestsList.vue';
 import hourglass from '@/assets/hourglass.png';
 import {getTimeRemaining} from '@/utils/common';
 import {NftIdType} from '@/components/smart/NftList.vue';
